library_platform/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
#Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-modeltranslation==0.18.11
django-rosetta==0.9.9
django-tinymce==3.6.1
django-taggit==4.0.0
django-filter==23.4
#django-allauth==0.57.0
celery==5.3.4
redis==5.0.1